Refactor HatproBinCombined class
